refactor: CQRS 적용 수정
