+++ src/main/java/com/backoffice/upjuyanolja/domain/accommodation/controller/AccommodationBackofficeController.java
@@ -11,7 +11,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -22,7 +21,6 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/backoffice-api/accommodations")
 @RequiredArgsConstructor
-@Validated
 public class AccommodationBackofficeController {
 
     private final AccommodationCommandUseCase accommodationCommandUseCase;
