+++ src/main/java/com/backoffice/upjuyanolja/domain/room/service/usecase/RoomQueryUseCase.java
@@ -1,6 +1,8 @@
 package com.backoffice.upjuyanolja.domain.room.service.usecase;
 
 import com.backoffice.upjuyanolja.domain.accommodation.entity.Accommodation;
+import com.backoffice.upjuyanolja.domain.room.dto.response.RoomInfoResponse;
+import com.backoffice.upjuyanolja.domain.room.dto.response.RoomPageResponse;
 import com.backoffice.upjuyanolja.domain.room.entity.Room;
 import com.backoffice.upjuyanolja.domain.room.entity.RoomImage;
 import com.backoffice.upjuyanolja.domain.room.entity.RoomOption;
@@ -15,39 +17,9 @@ import org.springframework.data.domain.Pageable;
 
 public interface RoomQueryUseCase {
 
-    Room saveRoom(Accommodation accommodation, Room request);
+    RoomPageResponse getRooms(long memberId, long accommodationId, Pageable pageable);
 
-    List<RoomImage> saveRoomImages(List<RoomImage> request);
+    RoomInfoResponse getRoom(long memberId, long roomId);
 
-    Page<Room> findAllByAccommodationId(long accommodationId, Pageable pageable);
-
-    List<Room> findByAccommodationId(long id);
-
-    Room findRoomById(long id);
-
-    boolean existsRoomByNameAndAccommodation(String name, Accommodation accommodation);
-
-    RoomImage findRoomImage(long roomImageId);
-
-    void deleteRoomImages(List<RoomImage> requests);
-
-    void saveRoomStock(RoomStock roomStock);
-
-    List<RoomStock> findStockByRoom(Room room);
-
-    List<RoomStock> findStocksByRoomAndDateAfter(Room room, LocalDate date);
-
-    @Builder
-    record RoomSaveRequest(
-        String name,
-        RoomPrice price,
-        int defaultCapacity,
-        int maxCapacity,
-        LocalTime checkInTime,
-        LocalTime checkOutTime,
-        int amount,
-        RoomOption option
-    ) {
-
-    }
+    List<RoomStock> getFilteredRoomStocksByDate(Room room, LocalDate startDate, LocalDate endDate);
 }
