+++ src/test/java/com/backoffice/upjuyanolja/domain/room/docs/RoomControllerDocsTest.java
@@ -29,6 +29,7 @@ import com.backoffice.upjuyanolja.domain.room.dto.response.RoomOptionResponse;
 import com.backoffice.upjuyanolja.domain.room.dto.response.RoomPageResponse;
 import com.backoffice.upjuyanolja.domain.room.dto.response.RoomsInfoResponse;
 import com.backoffice.upjuyanolja.domain.room.service.usecase.RoomCommandUseCase;
+import com.backoffice.upjuyanolja.domain.room.service.usecase.RoomQueryUseCase;
 import com.backoffice.upjuyanolja.global.security.SecurityUtil;
 import com.backoffice.upjuyanolja.global.util.RestDocsSupport;
 import java.util.List;
@@ -47,6 +48,9 @@ public class RoomControllerDocsTest extends RestDocsSupport {
 
     @MockBean
     private RoomCommandUseCase roomCommandUseCase;
+
+    @MockBean
+    private RoomQueryUseCase roomQueryUseCase;
 
     @MockBean
     private SecurityUtil securityUtil;
@@ -263,7 +267,7 @@ public class RoomControllerDocsTest extends RestDocsSupport {
             .build();
 
         given(securityUtil.getCurrentMemberId()).willReturn(1L);
-        given(roomCommandUseCase
+        given(roomQueryUseCase
             .getRooms(any(Long.TYPE), any(Long.TYPE), any(Pageable.class)))
             .willReturn(roomPageResponse);
 
@@ -338,7 +342,7 @@ public class RoomControllerDocsTest extends RestDocsSupport {
                 )
             ));
 
-        verify(roomCommandUseCase, times(1))
+        verify(roomQueryUseCase, times(1))
             .getRooms(any(Long.TYPE), any(Long.TYPE), any(Pageable.class));
     }
 
@@ -369,7 +373,7 @@ public class RoomControllerDocsTest extends RestDocsSupport {
             .build();
 
         given(securityUtil.getCurrentMemberId()).willReturn(1L);
-        given(roomCommandUseCase.getRoom(any(Long.TYPE), any(Long.TYPE)))
+        given(roomQueryUseCase.getRoom(any(Long.TYPE), any(Long.TYPE)))
             .willReturn(roomInfoResponse);
 
         // when then
@@ -418,7 +422,7 @@ public class RoomControllerDocsTest extends RestDocsSupport {
                 )
             ));
 
-        verify(roomCommandUseCase, times(1)).getRoom(any(Long.TYPE), any(Long.TYPE));
+        verify(roomQueryUseCase, times(1)).getRoom(any(Long.TYPE), any(Long.TYPE));
     }
 
     @Test
@@ -526,7 +530,7 @@ public class RoomControllerDocsTest extends RestDocsSupport {
                     fieldWithPath("deleteImages").description("삭제할 객실 이미지 배열"),
                     fieldWithPath("deleteImages[].id").description("객실 이미지 식별자")
                         .attributes(key("constraints")
-                            .value(roomImageAddDescriptions.descriptionsForProperty(
+                            .value(roomImageDeleteDescriptions.descriptionsForProperty(
                                 "id"))),
                     fieldWithPath("option").description("객실 옵션"),
                     fieldWithPath("option.airCondition").description("객실 에어컨 여부")
